Text Ectraction Single Region based
Remove specific exctraction regions and only have one: the row. Read the pdf step by step with small y steps. This removes
the need to check dimensions of the page and the columns and instead reads each page from top to bottom anyways.

Declare regex patterns per enums as a collection of all the patterns required to extract the required information.

In future, data cleansing is required to ensure the data quality of the input pdf is as expected