Atomic Lecture Daos
Instead of creating a single LectureDao for both blocks (first start, end, location etc.), create
a dao for each of the blocks. This doubles the entries, but allows for more granular search
eventually.

Fix: Adjust the regex pattern for the start dates as the dates of all lectures were mixed up
until now!