Basic, fixed  module handbook extraction
Extracts information from a single page of the module
handbook based on hardcoded line numbers to get an
understanding of the individual lines and their information.

The hardcoded line numbers need to be swapped out for a search-algorithm
that finds the correct lines in any text regardless of the line number
they appear in.

The model so far carries all values as strings. Data format transformations
are still required