Dynamically search content on first page
Instead of using hard coded line numbers to find
the content on the page, now use pre-defined headline
text to find the associated content for that headline.

This way, regardless of the page layout, the content can be found.

This is still problematic though in case the content is not on the
same page anymore