Move masther thesis regex to module regex
Split the previous change of moving the regex patterns
extension for the master thesis to the modules regex
patterns only. The master thesis does not appear
in the lectures list and thus should not be present there.

Additionally, this way, the patterns itself can carry more
precise names and the validation code is more readable