Update lecture code pattern for getLectureCode
The getLectureCodeFromPage method of the text stripper
was still based on a custom pattern to fetch the lecture
code either from the top or the bottom right of the page.

This pattern was not valid for master thesis modules and further
too complicated. The pattern is now based on the common
lecture code pattern + the modules master thesis pattern