Separate extraction logic
Move the extraction logic for each topic into a
dedicated service. This allows to add more
sophisticated extraction logic and more precise
unit testing. A manager controls the extraction by
having a dependency on each of the extraction services
and calling their logic.

The class was called manager to clearly separate it from
the two orchestrators which are even higher level and orchestrate
the entire flow of the program for either module or lecture