Match entire headline to find information
Previously, to find the information, the algorithm
checked each line against parts of the headlines in the pdf.
This showed the problem that, if such part appeared again
somewhere later on the page for the same module, the previously
extracted content would be overwritten.

Now, the headlines are checked entirely