Merge lectures and modules
Iterate over all lectures and try to find mathcing
modules via the lecture code. This should be done
with the lecture being as the main entity as the modules
can contain all lectures, even those that are not part of
this years lectures.

In case a match is not found, the resulting entity is flagged
to make it easier for the UI to react based on missing
information.

The results of the merge are written to JSON as well and thus
can be used as the new basis for the UI.