Read wirtschaft lectures pixel-based
Instead of reading the entire lecture row and
then matching the different parts by regex,
read the different parts as indipendent regions of the pdf.

For each row, draw a region for every field of the row
and extract the information directly. Thus, the information
for each field is more precise and possible regex matchings
dont need to worry about exceptions from other fields
as they only focus on the one field