Read all lectures area by area
Entirely switch the approach on how to read
lectures from the pdf. Now, the pdf is read line
by line and each line is separated into different
area, one for each field of a lecture.

This way, the extracted text always matched exactly one
field and thus can easily be matched against regex and
processed further.

This was necessary as the faculty wirtschaft lectures show
a lot of edge cases in comparison to the technik lectures.
Dates, locations, names, additions are all part of the name
of the lecture and thus global regex patterns for dates for
example did not match correctly. Further, single blocks and
single day blocks further increased this problem.

On top of that, a lot has been simplified. The validation logic
can now also happen independently from the extraction where before
both parts used the same regex and thus was kinda redundant.