Parallize lecture extraction
Extract both faculties and each of the pages
in parallel to speed up the process. This reduced
the runtime from 12 minutes down to around 4.

In some cases an exception about tff or end of file
is thrown, but does not seem to have any effect of the
outcome and thus is ignored so far