Extract additional information
Extract the additional information found in
the lecture name of wirtschaft lectures.
This includes the location, language information,
the study path, the semester and others.

This kind of information is then stored in a new field
of the lecture.

Addtionally, remove the pdf reader tests per page
as for each of them the entire document is extracted
leading to extremely long run times