Read wirtschaft modules
Refactor the code to accomodate another module
pdf file for the wirtschaft modules. This results
in two ModulePdfTextStripper and ModulePdfGroupingService
beans, one for each faculty as they work based on a state
and thus could not work on two pdfs in parallel.

Add the additional filename in the application.properties
that a config class uses to initialise the above mentioned beans.
Thus, the creation of the beans is nomore done via annotation
in order to clearly define which bean works on which file.