Extract multiline exam forms
Extract exam forms that are described in more
or less a proper sentence incl. whitespaces.
The previous logic split the input string on
whitespaces to fetch the different parts. This
does not work in case of whitespaces as part of
the information.

Instead, in this special case, split on the well-known
string 'Siehe Pruefungsordnung' which allows a simple
separation into two parts