Extract owner if no language available
Some wirtschaft lectures dont give information on
the language of the lecture, the "field" is just empty
and thus the line end directly after the owner.
The previous approach worked based on the last index -1
and thus cut some information of the owner.

The new approach first checks if language information is present
and then decides how to extract the owner based on the two
options