ModulePdfTextStripper returns list of pages with content
The text stripper now directly returns the PdPageTree as a list
and adds the content of each page to a wrapper object CasPdPage.
This way, the grouping service can work on the pages with text
already, which makes unit testing significantly easier as the
tests do no longer rely on a pdf or extended mocking but simply
creating the wrapper object to represent a PdPage and its content.

Additionally, obtain the lecture code of a page from the bottom
line instead the top line as the lecture code is present in each
and every page in the bottom line. This makes identifying which
page is currently being observed a lot easier