Replace german umlaute before processing
Replace the german umlaute before processing the modules
content. Validations on the leturing forms failed due
to the word Übung which is not matched by simply A-Z regex
but would require the specific unicode characters.

The reduce this maintainability problem, instead read the text
in the text stripper and directly pass it to the cleaning
service. The outcome does is then served as the original
module content, but without the german umlaute