Allow multidisciplinary competences lecture code
The lecture code for multidisciplinary competences (wirtschaft)
has one digit less than all other: XMX0101.
Thus a special regex was required to match it.

Instead of making the lecture code regex softer, I decided for
the strict second regex. This ensures that any normal lecture code
that was only partly read is not a false positive as the last digit
would be optional