Extract multiline module names
Some modules like W3M40007 have such long names that their
name incl. lecture code has a linebreak. This mixes the line
numbers up so that the english name was not properly extracted before.

Recognize when a module name comes in multiline to extract the
according line for the english name. To prevent the english name
to be overridden by the other case, only fill it when the field is
null. This way, its only populated once and not simply overridden
by the second check