Split additional info on command semicolon
The additional information passed in the name of the lecture of
wirtschaft lectures is either split by semicolon or by comma. In order
to ensure that the code can handle both and split accordingly, the input
string is first split by semicolon and each of the resulting pieces is
then split by comma. The resulting list will include all parts